implement and add base exceptions

+++ HR.LeaveManagementSystem/HR.LeaveManagementSystem.Application/Features/LeaveAllocationCQRS/Commands/DeleteLeaveAllocation/DeleteLeaveAllocationCommandHandler.cs
@@ -1,4 +1,6 @@
 ﻿using HR.LeaveManagementSystem.Application.Contracts.Persistence;
+using HR.LeaveManagementSystem.Application.Exeptions;
+using HR.LeaveManagementSystem.Domain;
 using MediatR;
 
 namespace HR.LeaveManagementSystem.Application.Features.LeaveAllocationCQRS.Commands.DeleteLeaveAllocation;
@@ -16,7 +18,12 @@ public class DeleteLeaveAllocationCommandHandler : IRequestHandler<DeleteLeaveAl
         // Retrieve domain entity object
         var leaveAllocationToDelete = await _leaveAllocationRepository.GetByIdAsync(request.Id);
 
-        // Verify that entity exists
+        // Verify that entity
+        if (leaveAllocationToDelete == null)
+        {
+            throw new NotFoundException(nameof(LeaveAllocation), request.Id);
+        }
+        
 
         // Remove from database
         await _leaveAllocationRepository.DeleteAsync(leaveAllocationToDelete);
