+++ HR.LeaveManagementSystem/HR.LeaveManagementSystem.Application/Features/LeaveRequestCQRS/Commands/DeleteLeaveRequest/DeleteLeaveRequestCommandHandler.cs
@@ -1,4 +1,6 @@
 ﻿using HR.LeaveManagementSystem.Application.Contracts.Persistence;
+using HR.LeaveManagementSystem.Application.Exeptions;
+using HR.LeaveManagementSystem.Domain;
 using MediatR;
 
 namespace HR.LeaveManagementSystem.Application.Features.LeaveRequestCQRS.Commands.DeleteLeaveRequest;
@@ -18,8 +20,12 @@ public class DeleteLeaveRequestCommandHandler : IRequestHandler<DeleteLeaveReque
         var leaveRequestToDelete = await _leaveRequestRepository.GetByIdAsync(request.Id);
 
         // Verify that object exists
+        if (leaveRequestToDelete == null)
+        {
+            throw new NotFoundException(nameof(LeaveRequest), request.Id);
+        }
 
-        // Remove from database
+            // Remove from database
         await _leaveRequestRepository.DeleteAsync(leaveRequestToDelete);
 
         // Return Unit value
